refactor: 로또를 구매하는 메서드를 Purchaser 클래스로 분리
